# docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Personal Style Network API",
    "version": "1.0.0",
    "description": "API cho ứng dụng Personal Style Network",
    "contact": {
      "name": "PSN Team",
      "email": "support@psn.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3052/v1/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/items": {
      "post": {
        "summary": "Tạo mới vật phẩm",
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "category"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "top",
                      "bottom",
                      "outerwear",
                      "dress",
                      "footwear",
                      "accessory",
                      "other"
                    ]
                  },
                  "subCategory": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  },
                  "pattern": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string"
                  },
                  "size": {
                    "type": "string"
                  },
                  "material": {
                    "type": "string"
                  },
                  "season": {
                    "type": "string",
                    "enum": [
                      "spring",
                      "summer",
                      "fall",
                      "winter",
                      "all"
                    ]
                  },
                  "occasion": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "imageUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo vật phẩm thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      },
      "get": {
        "summary": "Lấy danh sách vật phẩm của người dùng",
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Lọc theo danh mục"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Số lượng mỗi trang"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách vật phẩm"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "summary": "Lấy thông tin chi tiết vật phẩm",
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của vật phẩm"
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin chi tiết vật phẩm"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy vật phẩm"
          }
        }
      },
      "put": {
        "summary": "Cập nhật thông tin vật phẩm",
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của vật phẩm"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  },
                  "pattern": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật vật phẩm thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy vật phẩm"
          }
        }
      },
      "delete": {
        "summary": "Xóa vật phẩm",
        "tags": [
          "Items"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của vật phẩm"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa vật phẩm thành công"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy vật phẩm"
          }
        }
      }
    },
...
    "/outfits/{id}/items": {
      "post": {
        "summary": "Thêm vật phẩm vào trang phục",
        "tags": [
          "Outfits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của trang phục"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "itemId"
                ],
                "properties": {
                  "itemId": {
                    "type": "string"
                  },
                  "position": {
                    "type": "string",
                    "enum": [
                      "top",
                      "bottom",
                      "outer",
                      "accessory",
                      "footwear",
                      "other"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thêm vật phẩm vào trang phục thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy trang phục hoặc vật phẩm"
          }
        }
      }
    },
    "/outfits/{outfitId}/items/{itemId}": {
      "delete": {
        "summary": "Xóa vật phẩm khỏi trang phục",
        "tags": [
          "Outfits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "outfitId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của trang phục"
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của vật phẩm"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa vật phẩm khỏi trang phục thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy trang phục hoặc vật phẩm"
          }
        }
      }
    },
...
  },
...
}